% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simdata_mult}
\alias{simdata_mult}
\title{Creation of simulation data}
\usage{
simdata_mult(
  p,
  non_zero,
  sigma,
  coef_val1,
  coef_val2,
  coef_val3,
  coef_val4,
  struc_diff,
  corr,
  N
)
}
\arguments{
\item{p:}{number of covariates}

\item{non_zero:}{number of covariates that were not simulated as associated with the outcome}

\item{sigma:}{value of variance parameter for the construction of x}

\item{coef_val1:}{value of predictor that is simulated as truly associated with the first category of the response}

\item{coef_val2:}{value of predictor that is simulated as truly associated with the second and fifth category of the response}

\item{coef_val3:}{value of predictor that is simulated as truly associated with the third and sixth category of the response}

\item{coef_val4:}{value of predictor that is simulated as truly associated with the fourth and seventh category of the response}

\item{struc_diff:}{simulated structure of truly associated predictors}

\item{corr:}{simulated correlation between covariates}

\item{N:}{number of simulated number of observations}
}
\value{
Simulated data.frame for K=7 categories with x and y
}
\description{
Function for creation of simulation data
}
\details{
Constructs matrix of continuous covariates and response vector
}
\references{
Fuetterer, C., Nalenz, M., and Augustin, T. (2021).
Discriminative Power Lasso – Incorporating Discriminative Power of Genes into
Regularization-Based Variable Selection. Technical Report. Available under:
\url{https://epub.ub.uni-muenchen.de/91666/1/DPL_TR_2022_03.pdf}
}
